pyproject.toml
[tool.poetry]
name = "my-resume"
version = "0.1.0"
description = "Olivier Philippon's hand-crafted resume 🤓"
authors = ["Olivier Philippon <olivier@dunsap.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
django-tailwind-cli = "^2.8.0"
django-google-fonts = "^0.0.3"
playwright = "^1.40.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
ruff = "^0.1.9"
ipython = "^8.19.0"
django-browser-reload = "^1.12.1"
djlint = "^1.34.1"
pillow = "^10.3.0"


[tool.poetry.group.deployment.dependencies]
ghp-import = "^2.1.0"

[tool.mypy]
# @link https://mypy.readthedocs.io/en/stable/config_file.html#using-a-pyproject-toml
python_version = "3.12"
mypy_path = "myresume/"
[[tool.mypy.overrides]]
module = [
    "django.*",
    "djlint.*",
]
ignore_missing_imports = true

[tool.ruff]
exclude = [".venv"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]

[tool.ruff.lint.isort]
combine-as-imports = true
extra-standard-library = ["tomllib"]
known-first-party = ["myresume"]

[tool.djlint]
max_line_length = 88
profile = "django"
ignore = "H019"
blank_line_after_tag="load,extends,include,endblock"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
